requirements.txt
alembic==1.8.1
atomicwrites==1.4.1
attrs==22.1.0
Beaker==1.11.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.0
click==8.1.3
cryptography==38.0.1
defusedxml==0.7.1
flask-login==0.6.2
Flask-Migrate==3.1.0
flask-saml2==0.3.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
Flask==2.2.2
freezegun==1.2.2
future==0.18.2
html5lib==1.1
idna==3.4
iso8601==0.1.16
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==9.0.0
oic==1.4.0
pip==22.3
pluggy==0.7.1
py==1.11.0
pycparser==2.21
pycryptodomex==3.15.0
pyjwkest==1.4.2
pyOpenSSL==22.1.0
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
setuptools==65.5.0
signxml==2.10.1
six==1.16.0
SQLAlchemy==1.4.42
transliterate==1.10.2
typing-extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
WTForms==3.0.1
xmlunittest==0.5.0
